Refactoring the classes
