+++ build.gradle
@@ -6,7 +6,7 @@ plugins {
 }
 
 group = 'io.github.secretx33'
-version = '1.0.2'
+version = '1.0.3'
 
 sourceCompatibility = '1.8'
 targetCompatibility = '1.8'
@@ -24,7 +24,7 @@ repositories {
 }
 
 dependencies {
-    compileOnly 'org.spigotmc:spigot-api:1.16.3-R0.1-SNAPSHOT'
+    compileOnly 'org.spigotmc:spigot-api:1.16.5-R0.1-SNAPSHOT'
     compileOnly 'org.jetbrains:annotations:20.1.0'
     compileOnly 'com.google.errorprone:error_prone_annotations:2.4.0'
     compileOnly 'com.google.guava:guava:30.1-jre'
@@ -43,9 +43,6 @@ artifacts { archives shadowJar }
 shadowJar {
     archiveFileName = rootProject.name + '.jar'
     from('LICENSE').rename('(LICENSE)', '$1.txt')
-
-    /*relocate 'com.google.thirdparty', 'io.github.secretx.lib.thirdparty'
-    relocate 'com.google.common', 'io.github.secretx.lib.guava'*/
 }
 
 processResources {
