+++ src/main/java/io/github/secretx33/mobstatuschange/events/CreeperExplosionEvents.java
@@ -14,13 +14,13 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with MobStatusChange.  If not, see <https://www.gnu.org/licenses/>.
  */
-package io.github.secretx33.mobstatuschange.Events;
+package io.github.secretx33.mobstatuschange.events;
 
 import io.github.secretx33.mobstatuschange.Main;
-import io.github.secretx33.mobstatuschange.Config.Config;
-import io.github.secretx33.mobstatuschange.Config.Const;
-import io.github.secretx33.mobstatuschange.Entity.EntityAttributes;
-import io.github.secretx33.mobstatuschange.Utils.Utils;
+import io.github.secretx33.mobstatuschange.config.Config;
+import io.github.secretx33.mobstatuschange.config.Const;
+import io.github.secretx33.mobstatuschange.entity.EntityAttributes;
+import io.github.secretx33.mobstatuschange.utils.Utils;
 import org.bukkit.Bukkit;
 import org.bukkit.Material;
 import org.bukkit.Sound;
